fixed a compiler warn in Matrix4
git-svn-id: https://svn.intern.artcom.de/pro60/trunk/src@4302 f3027e6b-e3f5-0310-a8e1-edde99107618

+++ asl/math/Matrix4.h
@@ -435,7 +435,7 @@ namespace asl {
             if (zFar <= zNear) {
                 throw asl::Exception("Matrix4::makePerspective: zFar must be greater than zNear.", PLUS_FILE_LINE);
             }
-            Number f = cos(fovy/2.0) / sin(fovy/2.0);
+            Number f = float(cos(fovy/2.0) / sin(fovy/2.0));
             assign(f / aspect, 0, 0, 0,
                     0, f, 0, 0,
                     0, 0, (zFar + zNear) / (zNear - zFar), 2*zFar*zNear / (zNear - zFar),
